requirements.txt
django
Faker
pymemcache
djangorestframework
pytest
pytest-django